fix: wrong image path
Fixed an error where acquiring image references and patching them pointed to different locations. The acquisition of image references previsouly just stored all in a list, whether they were from the initContainers or not. During validation we iterated over this list, keeping each items index in mind so we can then patch the correct location in the received request. Problem is, that the initContainers and regular containers have differnt indizes in the request, while we stored all in the same list. The consequence of this was, should someone deploy a resource with valid initContainer and valid regular container, connaisseur will attempt to patch container indizes, that don't exist. Now each image will keep track of its own index and container_type ("containers" or "initContainers"), so when patching the request after validation, the correct container gets patched.